requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.40.0
pydantic>=2.12.0

# Supabase
supabase>=2.27.0

# Auth
python-jose>=3.3.0
pyjwt>=2.10.0

# Utils
python-dotenv>=1.0.0
email-validator>=2.0.0
